# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger Gentcmen API documentation",
        "contact": {
            "email": "nikitosnov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/client/development-ideas": {
            "get": {
                "tags": [
                    "Development ideas"
                ],
                "summary": "Fetch development ideas for client - auth user",
                "operationId": "Fetch development ideas for client - auth user",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Development ideas fetched successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Development ideas"
                ],
                "summary": "Create development idea",
                "operationId": "Create development idea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDevelopmentIdeaVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Development idea created successfully"
                    },
                    "404": {
                        "description": "Idea improvement not found"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/development-ideas": {
            "get": {
                "tags": [
                    "Development ideas"
                ],
                "summary": "Fetch development ideas for admin",
                "operationId": "Fetch development ideas for admin",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Development ideas fetched successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Development ideas"
                ],
                "summary": "Delete development ideas items",
                "operationId": "Delete development ideas items",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DestroySomeItemsVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Development ideas deleted successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/development-ideas/{idea}": {
            "delete": {
                "tags": [
                    "Development ideas"
                ],
                "summary": "Delete development idea by id",
                "operationId": "Delete development idea by id",
                "parameters": [
                    {
                        "name": "idea",
                        "in": "path",
                        "description": "Idea id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Development idea deleted successfully"
                    },
                    "404": {
                        "description": "Development idea not found"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Development ideas"
                ],
                "summary": "Patch update development-idea",
                "operationId": "Patch update development-idea",
                "parameters": [
                    {
                        "name": "idea",
                        "in": "path",
                        "description": "idea id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "title": "Status of faq",
                                        "description": "Status",
                                        "example": "received"
                                    },
                                    "message": {
                                        "title": "message of faq",
                                        "description": "message",
                                        "example": "some message"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Development-idea updated successfully"
                    },
                    "404": {
                        "description": "Development-idea not found"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/client/development-ideas/{idea}": {
            "put": {
                "tags": [
                    "Development ideas"
                ],
                "summary": "Update development-idea",
                "operationId": "Update development-idea",
                "parameters": [
                    {
                        "name": "idea",
                        "in": "path",
                        "description": "Idea id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDevelopmentIdeaVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Development idea updated successfully"
                    },
                    "404": {
                        "description": "Development idea not found"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateDevelopmentIdeaVirtualBody": {
                "title": "Create development idea example",
                "description": "Create development idea body data",
                "required": [
                    "user_name",
                    "email",
                    "idea"
                ],
                "properties": {
                    "user_name": {
                        "title": "user_name",
                        "description": "user name who create idea improvement",
                        "type": "string",
                        "example": "user name"
                    },
                    "email": {
                        "title": "email",
                        "description": "user email who create idea improvement",
                        "type": "string",
                        "example": "test@gmail.com"
                    },
                    "idea": {
                        "title": "idea",
                        "description": "idea content",
                        "type": "string",
                        "example": "idea content goes here"
                    }
                },
                "type": "object"
            },
...
            "DestroySomeItemsVirtualBody": {
                "title": "Destroy some items example",
                "description": "Destroy some items body data",
                "required": [
                    "ids"
                ],
                "properties": {
                    "ids": {
                        "title": "ids",
                        "description": "Identifiers that should be deleted",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}